Support parsing title from snippets
Displays as a centered line at the top of the screen. I long line that
wraps will probably break the font size of the content element. The
`plugin.js` script from trmnl starts the text at the started size
("xxlarge" here) and cranks it down until the content element fits in the
"data-value-fit-max-height". If the element never gets that small it'll
make the text very small (I think you can set a min size actually).
Anyway, I had to adjust the max size with the title element here, not
sure why. It was working fine at 300px before...