Fix max content element height calculation
yay for fixed size viewports.

These sizes are the max size of the content element with and without one
line each of the title and attribution elements.

The plugin js uses the max size to know when to stop reducing the font
size at part of its "fit value" feature.