Adjust context max height again
The `fitValue()` algorithm started making the text really small
overnight because it couldn't get the content element to get small
enough. I'm not sure what changed, but presumably trmnl updated
something and made more space available.

For this setup the content element will stop shrinking when no elements
are overflowing the container anymore. It's got stretch-y so it'll
automatically take up all the rest of the space, even if the text in it
is very small. So a more precise text resizing algorithm for this case
would be to stop resizing the text one the element stops changing size.
That seems like a harmless check for any setup actually?

Anyway, for now do some math to calculate the available space and re-run
`fitValue()` with probably max-size - the 10px gap set. So the loop
should stop around about the target size of the element (the minumum
size I guess).

Hopefully the 250ms timeout isn't too short. This needs to run after
trml's onload hook.